+++ src/main/java/com/eleks/academy/whoami/configuration/WebSecurity.java
@@ -1,12 +1,10 @@
 package com.eleks.academy.whoami.configuration;
 
-import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
 @Configuration
-@RequiredArgsConstructor
 public class WebSecurity {
 
     @Bean
