+++ src/main/java/com/eleks/academy/whoami/core/impl/PersistentGame.java
@@ -11,7 +11,6 @@ import com.eleks.academy.whoami.model.request.CharacterSuggestion;
 import com.eleks.academy.whoami.model.request.Message;
 import com.eleks.academy.whoami.model.response.PlayerDetails;
 import com.eleks.academy.whoami.model.response.TurnDetails;
-import com.eleks.academy.whoami.repository.HistoryChat;
 
 import java.time.Instant;
 import java.util.*;
