+++ src/test/resources/models/game/answer-request.json
@@ -1,3 +1,3 @@
 {
-	"message": "Question"
+  "message": "Question"
 }
