+++ src/test/resources/models/game/ask-question-request.json
@@ -1,3 +1,3 @@
 {
-	"message": "Question"
+  "message": "Question"
 }
