+++ src/test/resources/models/game/create-game-request.json
@@ -1,3 +1,3 @@
 {
-	"maxPlayers": 4
+  "maxPlayers": 4
 }
