+++ src/test/resources/models/game/enroll-player-response.json
@@ -1,4 +1,4 @@
 {
-	"name": "Example",
-	"character": "Batman"
+  "name": "Example",
+  "character": "Batman"
 }
