+++ src/test/resources/models/game/get-available-games-response.json
@@ -1,7 +1,7 @@
 [
-	{
-		"id": "1234-UUID",
-		"status": "WaitingForPlayers",
-		"playersInGame": "3/4"
-	}
+  {
+    "id": "1234-UUID",
+    "status": "WaitingForPlayers",
+    "playersInGame": "3/4"
+  }
 ]
