+++ src/test/resources/models/game/get-current-turn-response.json
@@ -1,16 +1,16 @@
 {
-	"currentPlayer": {
-		"name": "Example",
-		"character": "Batman"
-	},
-	"players": [
-		{
-			"player": {
-				"name": "Example",
-				"character": "Batman"
-			},
-			"answer": "NO",
-			"state": "NOT_READY"
-		}
-	]
+  "currentPlayer": {
+    "name": "Example",
+    "character": "Batman"
+  },
+  "players": [
+    {
+      "player": {
+        "name": "Example",
+        "character": "Batman"
+      },
+      "answer": "NO",
+      "state": "NOT_READY"
+    }
+  ]
 }
