+++ src/test/resources/models/game/get-game-by-id-response.json
@@ -1,15 +1,15 @@
 {
-	"id": "1234-Uid",
-	"status": "WaitingForPlayers",
-	"currentTurn": "1",
-	"players": [
-		{
-			"player": {
-				"name": "Example",
-				"character": "Batman"
-			},
-			"answer": "NO",
-			"state": "NOT_READY"
-		}
-	]
+  "id": "1234-Uid",
+  "status": "WaitingForPlayers",
+  "currentTurn": "1",
+  "players": [
+    {
+      "player": {
+        "name": "Example",
+        "character": "Batman"
+      },
+      "answer": "NO",
+      "state": "NOT_READY"
+    }
+  ]
 }
