+++ src/test/resources/models/game/guess-request.json
@@ -1,3 +1,3 @@
 {
-	"message": "Question"
+  "message": "Question"
 }
