+++ src/test/resources/models/game/suggest-character-request.json
@@ -1,3 +1,3 @@
 {
-	"character": "Batman"
+  "character": "Batman"
 }
